+++ src/autoadd_bindir/__init__.py
@@ -32,29 +32,3 @@ def get_bindir() -> pl.Path:
     """Return the bindir from the isolated virtual environment."""
     bin_dir = pl.Path(sys.prefix) / "bin"
     return bin_dir
-
-
-def cancel() -> None:  # pragma: no cover
-    """No-op function that can be used the cancel a registered entrypoint.
-
-    Imagine you have multiple sitecustomize-entrypoints. If these entrypoints
-    are registered via third-party packages, you cannot control the order of execution.
-
-    Now suppose some of these entrypoints need an environment-variable that first need to be set
-    by ``autoadd_bindir`` needs to be executed before the others
-
-    entrypoint 1:  foo.needs_envvar:bar
-    entrypoint 2:  autoadd_bindir.entrypoint:autoadd_bindir
-
-    in your project's pyproject.toml:
-
-    [tool.poetry.plugins."sitecustomize"]
-
-    # cancel the first registration using the original name
-    autoadd_bindir = "autoadd_bindir.entrypoint:cancel"
-
-    # re-register the same function under different name
-    zz_autoadd_bindir = "autoadd_bindir.entrypoint:autoadd_bindir"
-
-    """
-    pass  # pylint: disable=unnecessary-pass