+++ docs/changes.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 
 
+## Unreleased (YYYY-MM-DD)
+
+- Remove ``autoadd_bindir.cancel``-function
+
+- Depend on ``sitecustomize-entrypoints>=1.1.0``
 
 
 ## 1.0.4 (2023-03-29)